session module refactoring
